Reuse inflater, rename fragment
